+++ src/common/api/index.ts
@@ -1,9 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import {
-  MagicLinkParams,
-  postV1MagicLink,
-  postV1SandboxEmploymentsEmploymentIdRiskReserveProofOfPaymentsApprove,
-} from '@/src/client';
+import { MagicLinkParams, postV1MagicLink } from '@/src/client';
 import { useClient } from '@/src/context';
 
 import { Client } from '@/src/client/client';
@@ -19,25 +15,6 @@ export const useMagicLink = () => {
         },
         body: params,
       });
-    },
-  });
-};
-
-export const useSandboxApproveRiskReservePayment = () => {
-  const { client } = useClient();
-  return useMutation({
-    mutationFn: (employmentId: string) => {
-      return postV1SandboxEmploymentsEmploymentIdRiskReserveProofOfPaymentsApprove(
-        {
-          client: client as Client,
-          headers: {
-            Authorization: ``,
-          },
-          path: {
-            employment_id: employmentId,
-          },
-        },
-      );
     },
   });
 };
